also rename the identity cache package
