+++ internal/identity/cache.go
@@ -1,4 +1,4 @@
-package identitycache
+package identity
 
 import (
 	"context"
