+++ internal/identity/cache_test.go
@@ -1,4 +1,4 @@
-package identitycache
+package identity
 
 import (
 	"encoding/binary"
